profiles: Bug 524364 - libglade -> gtkbuilder
Finally remove glade from GNOME Media.

I figure out that the profiles/glade directory contain a glade plugin,
not for others to use in their own UI, but to be able to modify the
GNOME Media UI gnome-audio-profile-edit.ui itself.

This is quite important to have for now. However, A saner solution
would be to fill a regular dialog, thus avoiding a custom dialog and
its glade registration.